4.20-4.23

+++ osa4/blogilista/controllers/blogs.js
@@ -3,6 +3,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog.js')
 const User = require('../models/user.js')
 const jwt = require('jsonwebtoken')
+const userExtractor = require("../utils/middleware").userExtractor
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user',{username:1,name:1})
@@ -16,35 +17,27 @@ blogsRouter.get('/:id', async(request, response) => {
   })
 
 
-  const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')){
-      return authorization.replace('Bearer ','')
-    }
-    return null
+blogsRouter.post('/', userExtractor, async (request, response) => {
+  if(!request.token){
+    response.status(401).json({error: 'token invalid'})
   }
-
-blogsRouter.post('/', async (request, response) => {
-  const decodedToken = jwt.verify(getTokenFrom(request),process.env.SECRET)
-  console.log(decodedToken)
+  const decodedToken = jwt.verify(request.token,process.env.SECRET)
   if(!decodedToken.id){
     return response.status(401).json({error: 'token invalid'})
   }
-  console.log(request.body)
   const body=request.body
  // const blog = new Blog(request.body)
- const user = await User.findById(decodedToken.id)
-  const blog = new Blog({title: body.title,user:user._id, author:body.author, url:body.url, likes:body.likes})
+  const blog = new Blog({title: body.title,user:request.user._id, author:body.author, url:body.url, likes:body.likes})
 
   const savedBlog = await blog.save()
 
-  user.blogs = user.blogs.concat(savedBlog._id)
-  await user.save()
+  request.user.blogs = request.user.blogs.concat(savedBlog._id)
+  await request.user.save()
   response.status(201).json(savedBlog)
 
 })
 
-blogsRouter.put('/:id',async(request,response)=>{
+blogsRouter.put('/:id', userExtractor, async(request,response)=>{
   const body = request.body
   const blog = {
     title: body.title,
@@ -57,10 +50,24 @@ blogsRouter.put('/:id',async(request,response)=>{
   response.json(res)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  //console.log(request.params)
-  const res = await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).json(res)
+blogsRouter.delete('/:id',userExtractor, async (request, response) => {
+  //const user = request.user
+  //console.log(user)
+  const decodedToken = jwt.verify(request.token,process.env.SECRET)
+  if(!decodedToken.id){
+    return response.status(401).json({error: 'token invalid'})
+  }
+  const blog = await Blog.findById(request.params.id)
+  if(blog===null){ response.status(404).json({error:"blog not found"}
+  )}
+
+  else if ( blog.user.toString() === request.user._id.toString() ){
+    const blogDeleted = await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).json(blogDeleted)
+  }
+  else{
+    response.status(403).json({"error":"wrong user"})
+  }
 })
 
 module.exports = blogsRouter
